+++ RTv1/src/vector.c
@@ -12,13 +12,13 @@ t_pos *vectorSub(t_pos *v1, t_pos *v2)
 }
 
 /* Multiply two vectors and return the resulting scalar (dot product) */
-double vectorDot(t_pos *v1, t_pos *v2)
+float vectorDot(t_pos *v1, t_pos *v2)
 {
     return v1->x * v2->x + v1->y * v2->y + v1->z * v2->z;
 }
 
 /* Calculate Vector x Scalar and return resulting Vector*/
-t_pos *vectorScale(double c, t_pos *v)
+t_pos *vectorScale(float c, t_pos *v)
 {
     t_pos *result;
